Clean
The data is now being rendered locally as it should be. Typing the command 'npm run dev' in the repo root seemed to fix the issue. Now, the data is rendered locally when 'node server.js' is ran in the api directory.

Let's see if the data renders on render though...